4th commit. Error correction
Added the function of jumping (Space) and running (Shift) for the player.
Fixed the work of the turret. It shot both mobs and the player in the legs.

After the destruction of the mob, it did not look for a new target, but simply fired in an unknown direction.

Fixed work with the HP of the player and mobs. Now mobs, like the player, disappear after such as the end of HP